Restructured user module
