--- backend/src/main/java/pl/lodz/p/it/opinioncollector/user/UserType.java
@@ -1,6 +1,0 @@
-package pl.lodz.p.it.opinioncollector.user;
-
-public enum UserType {
-    USER,
-    ADMIN
-}
