+++ backend/src/main/java/pl/lodz/p/it/opinioncollector/userModule/auth/AuthController.java
@@ -1,4 +1,4 @@
-package pl.lodz.p.it.opinioncollector.auth;
+package pl.lodz.p.it.opinioncollector.userModule.auth;
 
 
 import lombok.RequiredArgsConstructor;
