+++ backend/src/main/java/pl/lodz/p/it/opinioncollector/userModule/auth/AuthManager.java
@@ -1,12 +1,13 @@
-package pl.lodz.p.it.opinioncollector.user;
+package pl.lodz.p.it.opinioncollector.userModule.auth;
 
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import pl.lodz.p.it.opinioncollector.userModule.user.UserRepository;
 
 @Service
 @RequiredArgsConstructor
-public class UserManager {
+public class AuthManager {
 
     private final UserRepository userRepository;
