+++ backend/src/main/java/pl/lodz/p/it/opinioncollector/userModule/auth/LoginDTO.java
@@ -1,0 +1,17 @@
+package pl.lodz.p.it.opinioncollector.userModule.auth;
+
+import jakarta.validation.constraints.NotNull;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class LoginDTO {
+    @NotNull
+    String email;
+
+    @NotNull
+    String password;
+}
