+++ backend/src/main/java/pl/lodz/p/it/opinioncollector/userModule/user/UserType.java
@@ -1,0 +1,6 @@
+package pl.lodz.p.it.opinioncollector.userModule.user;
+
+public enum UserType {
+    USER,
+    ADMIN
+}
